Removed extra offline banners and moved alert to alert/containers
